requirements.txt
python-telegram-bot==20.7
pywin32
numpy
opencv-python